refactor: home remaining large inline unit tests
Move protocol, cache, state, CTC, features, VAD windows, load-files,
sizing, token-format, manifest, error, config, metrics, middleware,
server, and Core ML encoder-session tests next to their impls.
