+++ CHANGELOG.md
@@ -30,10 +30,11 @@ were released without a git tag, so their headings carry no compare link.
 - `model` download/identity split into `model/{progress,variant,download}.rs`.
   Public `gigastt_core::model::*` paths are unchanged. CLI parse tests live
   next to `main.rs`; bind-gate tests live next to `serve.rs`.
-- Audio, engine, export, decode, punctuation, VAD, FFI, model, HTTP,
-  wordpiece, stream, quantize, batch, boot, serve, bias, pool, rate-limit,
-  and Core ML bridge unit tests live next to their modules. Jobs unit
-  tests are split by store / queue / events.
+- Unit tests live next to their modules (audio, engine, export, decode,
+  punctuation, VAD, FFI, model, HTTP, wordpiece, stream, quantize, batch,
+  boot, serve, bias, pool, rate-limit, protocol, cache, state, CTC,
+  features, metrics, and the rest). Jobs tests are split by store / queue
+  / events. WordPiece Unicode category tables live in `wordpiece/unicode.rs`.
 
 ### Fixed
